requirements.txt
agentmail>=0.0.19
agentmail-toolkit>=0.1.8
flask>=3.1.0
ngrok>=1.4.0
openai-agents>=0.0.9
python-dotenv

